Defer setting up of the playerControlsView until it's enabled
Summary:
We pass in `controls={false}`, and so we never want the controls view displayed.
Incidentally, with react-native 0.60.0 we get a crash at line 291, because `R.id.exo_play_pause_contrainer` isn't defined.
I started down the rabbit hole of figuring out why that is, but given that we don't actually use the controls view,
this ended up being a short + sweet fix.
It's still probably a good idea to figure out what's going on with the resources at some point.

Test plan:
Build & run our app with this change (on the 0.60.0 branch), and see that videos play! And it doesn't crash!